Removed optional slice as first arg from all slices.
It wasn't being used. It was intended to be used as (-> slice1 slice2) in the
body of a slice which would just be a regular function before I decided to
merge slices with the slice macro. I may be forgetting somewhere it's useful
but for simple use of combining multiple slices w/o giving them a name the
slices function works.